Move clear_collections to own commands module
