+++ plex_trakt_sync/cli.py
@@ -1,5 +1,5 @@
 import click
-from plex_trakt_sync.clear_trakt_collections import clear_trakt_collections
+from plex_trakt_sync.commands.clear_collections import clear_collections
 from plex_trakt_sync.commands.sync import sync
 
 
@@ -11,19 +11,6 @@ def cli(ctx):
     """
     if not ctx.invoked_subcommand:
         sync()
-
-
-@click.command()
-@click.option('--confirm', is_flag=True, help='Confirm the dangerous action')
-def clear_collections(confirm):
-    """
-    Clear Movies and Shows collections in Trakt
-    """
-
-    if not confirm:
-        click.echo('You need to pass --confirm option to proceed')
-        return
-    clear_trakt_collections()
 
 
 cli.add_command(sync)
